endpoints for create a new room and retrieve all the rooms.
not yet: add a user to a room; start a game;

+++ src/main/java/ch/uzh/ifi/hase/soprafs23/game/Lobby.java
@@ -30,7 +30,7 @@ public class Lobby {
         return room;
     }
 
-    public Room getRoomByRoomId(int roomId) { return roomMap.get(roomId); }
+    public Room getRoomByRoomId(Long roomId) { return roomMap.get(roomId); }
 
     public HashMap<Integer, Room> getRooms() { return this.roomMap; }
